[improvement] PasswordInput: jsx (#2611)
